fix: split steps into individual files
